- Clarified help   - pheader -c accepts more aliases of the same parameters   - Potential problem when opening multiple input files resulting in a segmentation fault   - A specific error is given if data in timer format is being tried to be read in   - Improved the reading of header information for older PSRFITS versions.   - Improved read speed of psrfits files.   - Implemented the inferno colour map   - Added the -cont option in pplot   - Added the -titlescale option in pspecFig   - Added the -labels, -lw and -vp options in pdist.   - Made the code compilable on MAC OS X   - Made some changes to code to reduce compiler warnings. This shouldn't impact the behaviour, except in an unsignificant way related to rounding errors etc.   - Increased efficiency of some pgplot related code
Patrick

+++ src/lib/psrio_sigproc.c
@@ -357,7 +357,7 @@ int readSigprocHeader(datafile_definition *datafile, verbose_definition verbose)
   datasize -= curpos;
   fseeko(datafile->fptr, curpos, SEEK_SET);
   if(verbose.debug) {
-    printf("DEBUG: data size=%ld byte\n", datasize);
+    printf("DEBUG: data size=%ld byte\n", (long)datasize);
   }
   if(datafile->isFolded && datafile->NrBins <= 0) {
     printwarning(verbose.debug, "WARNING readSigprocHeader: For folded data the number of bins should be defined in header. Assume the data is not folded");
